+++ indico/modules/receipts/client/js/templates/Previewer.jsx
@@ -15,8 +15,7 @@ import {Translate} from 'indico/react/i18n';
 import {indicoAxios, handleAxiosError} from 'indico/utils/axios';
 import {makeAsyncDebounce} from 'indico/utils/debounce';
 
-import TemplateParameterEditor from '../printing/TemplateParameterEditor';
-import {getDefaultFieldValue} from '../printing/util';
+import TemplateParameterEditor, {getDefaultFieldValue} from '../printing/TemplateParameterEditor';
 
 import 'react-pdf/dist/esm/Page/AnnotationLayer.css';
 import 'react-pdf/dist/esm/Page/TextLayer.css';
